Speed up escalation auditor: - use raw escalation snapshot instead of serialized one
